Tidy FileStatusList up
No functional changes

* Reorder `FileStatusList` members
* Extracted `GitItemStatusWithParent` into own file
* Renamed `AlwaysRevisionGroups` to `GroupByRevision`, added getter
