+++ GitUI/UserControls/FileStatusList.cs
@@ -27,18 +27,23 @@ namespace GitUI
     public sealed partial class FileStatusList : GitModuleControl
     {
         private static readonly TimeSpan SelectedIndexChangeThrottleDuration = TimeSpan.FromMilliseconds(50);
-
         private readonly TranslationString _diffWithParent = new TranslationString("Diff with:");
         public readonly TranslationString CombinedDiff = new TranslationString("Combined Diff");
-
         private readonly IGitRevisionTester _revisionTester;
         private readonly IFullPathResolver _fullPathResolver;
-
-        private bool _alwaysRevisionGroups;
+        private int _nextIndexToSelect = -1;
+        private bool _groupByRevision;
+        private bool _enableSelectedIndexChangeEvent = true;
         private ToolStripItem _openSubmoduleMenuItem;
+        private Rectangle _dragBoxFromMouseDown;
+        private IReadOnlyList<(GitRevision revision, IReadOnlyList<GitItemStatus> statuses)> _itemsWithParent = Array.Empty<(GitRevision, IReadOnlyList<GitItemStatus>)>();
         [CanBeNull] private IDisposable _selectedIndexChangeSubscription;
 
-        public DescribeRevisionDelegate DescribeRevision { get; set; }
+        public event EventHandler SelectedIndexChanged;
+        public event EventHandler DataSourceChanged;
+
+        public new event EventHandler DoubleClick;
+        public new event KeyEventHandler KeyDown;
 
         public FileStatusList()
         {
@@ -103,41 +108,37 @@ namespace GitUI
             }
         }
 
-        public bool AlwaysRevisionGroups
+        // Properties
+
+        [Browsable(false)]
+        public IEnumerable<GitItemStatus> AllItems => FileStatusListView.Items.Cast<ListViewItem>().Select(selectedItem => (GitItemStatus)selectedItem.Tag);
+
+        public int AllItemsCount => FileStatusListView.Items.Count;
+
+        public override ContextMenu ContextMenu
         {
-            set => _alwaysRevisionGroups = value;
+            get => FileStatusListView.ContextMenu;
+            set => FileStatusListView.ContextMenu = value;
         }
 
-        private void CreateOpenSubmoduleMenuItem()
+        public override ContextMenuStrip ContextMenuStrip
         {
-            _openSubmoduleMenuItem = new ToolStripMenuItem
+            get { return FileStatusListView.ContextMenuStrip; }
+            set
             {
-                Name = "openSubmoduleMenuItem",
-                Tag = "1",
-                Text = "Open with Git Extensions",
-                Image = Images.GitExtensionsLogo16
-            };
-            _openSubmoduleMenuItem.Click += (s, ea) => { ThreadHelper.JoinableTaskFactory.RunAsync(() => OpenSubmoduleAsync()); };
+                FileStatusListView.ContextMenuStrip = value;
+                if (FileStatusListView.ContextMenuStrip != null)
+                {
+                    FileStatusListView.ContextMenuStrip.Opening += FileStatusListView_ContextMenu_Opening;
+                }
+            }
         }
 
-        protected override void DisposeCustomResources()
-        {
-            _selectedIndexChangeSubscription?.Dispose();
-        }
+        [DesignerSerializationVisibility(DesignerSerializationVisibility.Hidden)]
+        [Browsable(false)]
+        public DescribeRevisionDelegate DescribeRevision { get; set; }
 
-        private bool _enableSelectedIndexChangeEvent = true;
-
-        public void SetSelectedIndex(int idx, bool notify)
-        {
-            _enableSelectedIndexChangeEvent = notify;
-            SelectedIndex = idx;
-            _enableSelectedIndexChangeEvent = true;
-        }
-
-        public void SetNoFilesText(string text)
-        {
-            NoFiles.Text = text;
-        }
+        public bool FilterFocused => FilterComboBox.Focused;
 
         public bool FilterVisible
         {
@@ -150,9 +151,199 @@ namespace GitUI
             }
         }
 
-        public bool FilterFocused => FilterComboBox.Focused;
-
         public override bool Focused => FileStatusListView.Focused;
+
+        [DesignerSerializationVisibility(DesignerSerializationVisibility.Hidden)]
+        [Browsable(false)]
+        public IReadOnlyList<GitItemStatus> GitItemFilteredStatuses
+        {
+            get
+            {
+                var result = new List<GitItemStatus>(FileStatusListView.Items.Count);
+
+                foreach (ListViewItem listViewItem in FileStatusListView.Items)
+                {
+                    result.Add(listViewItem.Tag as GitItemStatus);
+                }
+
+                return result;
+            }
+        }
+
+        [DesignerSerializationVisibility(DesignerSerializationVisibility.Hidden)]
+        [Browsable(false)]
+        public IReadOnlyList<GitItemStatus> GitItemStatuses
+        {
+            get
+            {
+                return GitItemStatusesWithParents?.SelectMany(tuple => tuple.statuses).ToList()
+                       ?? (IReadOnlyList<GitItemStatus>)Array.Empty<GitItemStatus>();
+            }
+        }
+
+        // Parents is used as the "first selected" (not always the parent) for GitItemStatus
+        [DesignerSerializationVisibility(DesignerSerializationVisibility.Hidden)]
+        [Browsable(false)]
+        public IReadOnlyList<(GitRevision revision, IReadOnlyList<GitItemStatus> statuses)> GitItemStatusesWithParents
+        {
+            get { return _itemsWithParent; }
+            private set
+            {
+                _itemsWithParent = value ?? throw new ArgumentNullException(nameof(value));
+                UpdateFileStatusListView();
+            }
+        }
+
+        public bool GroupByRevision
+        {
+            get => _groupByRevision;
+            set => _groupByRevision = value;
+        }
+
+        [Browsable(false)]
+        [DefaultValue(true)]
+        public bool IsEmpty => GitItemStatuses == null || !GitItemStatuses.Any();
+
+        /// <summary>
+        /// Gets or sets the revision.
+        /// </summary>
+        /// <value>The revision.</value>
+        [DesignerSerializationVisibility(DesignerSerializationVisibility.Hidden)]
+        [Browsable(false)]
+        public GitRevision Revision { get; set; }
+
+        [DesignerSerializationVisibility(DesignerSerializationVisibility.Hidden)]
+        [Browsable(false)]
+        public int SelectedIndex
+        {
+            get
+            {
+                foreach (int i in FileStatusListView.SelectedIndices)
+                {
+                    return i;
+                }
+
+                return -1;
+            }
+            set
+            {
+                ClearSelected();
+                if (value >= 0)
+                {
+                    FileStatusListView.Items[value].Selected = true;
+                    FileStatusListView.Items[value].Focused = true;
+                    FileStatusListView.Items[value].EnsureVisible();
+                }
+            }
+        }
+
+        [CanBeNull]
+        [DesignerSerializationVisibility(DesignerSerializationVisibility.Hidden)]
+        [Browsable(false)]
+        public GitItemStatus SelectedItem
+        {
+            get { return SelectedItems.FirstOrDefault(); }
+            set
+            {
+                ClearSelected();
+                if (value == null)
+                {
+                    return;
+                }
+
+                ListViewItem newSelected = null;
+                foreach (ListViewItem item in FileStatusListView.Items)
+                {
+                    if (value.CompareTo(item.Tag as GitItemStatus) == 0)
+                    {
+                        if (newSelected == null)
+                        {
+                            newSelected = item;
+                        }
+                        else if (item.Tag == value)
+                        {
+                            newSelected = item;
+                            break;
+                        }
+                    }
+                }
+
+                if (newSelected != null)
+                {
+                    newSelected.Selected = true;
+                    newSelected.EnsureVisible();
+                }
+            }
+        }
+
+        [CanBeNull]
+        [DesignerSerializationVisibility(DesignerSerializationVisibility.Hidden)]
+        [Browsable(false)]
+        public GitRevision SelectedItemParent => SelectedItemParents.FirstOrDefault();
+
+        [DesignerSerializationVisibility(DesignerSerializationVisibility.Hidden)]
+        [Browsable(false)]
+        public IEnumerable<GitRevision> SelectedItemParents
+        {
+            get
+            {
+                return FileStatusListView.SelectedItems.Cast<ListViewItem>()
+                    .Where(i => i.Group?.Tag is GitRevision)
+                    .Select(i => (GitRevision)i.Group.Tag);
+            }
+        }
+
+        [DesignerSerializationVisibility(DesignerSerializationVisibility.Hidden)]
+        [Browsable(false)]
+        public IEnumerable<GitItemStatus> SelectedItems
+        {
+            get { return FileStatusListView.SelectedItems.Cast<ListViewItem>().Select(i => (GitItemStatus)i.Tag); }
+            set
+            {
+                ClearSelected();
+                if (value == null)
+                {
+                    return;
+                }
+
+                foreach (var item in FileStatusListView.Items.Cast<ListViewItem>()
+                    .Where(i => value.Contains(i.Tag as GitItemStatus)))
+                {
+                    item.Selected = true;
+                }
+
+                var first = FileStatusListView.SelectedItems.Cast<ListViewItem>().FirstOrDefault(x => x.Selected);
+                first?.EnsureVisible();
+                StoreNextIndexToSelect();
+            }
+        }
+
+        [DesignerSerializationVisibility(DesignerSerializationVisibility.Hidden)]
+        [Browsable(false)]
+        public IEnumerable<GitItemStatusWithParent> SelectedItemsWithParent
+        {
+            get
+            {
+                return FileStatusListView.SelectedItems.Cast<ListViewItem>()
+                    .Where(i => i.Group?.Tag is GitRevision)
+                    .Select(i => new GitItemStatusWithParent((GitRevision)i.Group.Tag, i.Tag as GitItemStatus));
+            }
+        }
+
+        [DefaultValue(true)]
+        public bool SelectFirstItemOnSetItems { get; set; }
+
+        public int UnfilteredItemsCount => GitItemStatusesWithParents?.Sum(tuple => tuple.statuses.Count) ?? 0;
+
+        // Public methods
+
+        public void ClearSelected()
+        {
+            foreach (ListViewItem item in FileStatusListView.SelectedItems)
+            {
+                item.Selected = false;
+            }
+        }
 
         public new void Focus()
         {
@@ -302,367 +493,46 @@ namespace GitUI
             }
         }
 
-        private void FileStatusListView_DrawSubItem(object sender, DrawListViewSubItemEventArgs e)
+        public void SelectAll()
         {
-            if (!(e.Item?.Tag is GitItemStatus gitItemStatus))
+            try
+            {
+                SuspendLayout();
+
+                var itemCount = AllItems.Count();
+
+                for (var i = 0; i < itemCount; i++)
+                {
+                    FileStatusListView.Items[i].Selected = true;
+                    i++;
+                }
+            }
+            finally
+            {
+                ResumeLayout(true);
+            }
+        }
+
+        public void SelectFirstVisibleItem()
+        {
+            if (FileStatusListView.Items.Count == 0)
             {
                 return;
             }
 
-            var imageWidth = 0;
-            if (e.Item.ImageList != null && e.Item.ImageIndex != -1)
+            var group = FileStatusListView.Groups.Cast<ListViewGroup>().FirstOrDefault(gr => gr.Items.Count > 0);
+            if (group != null)
             {
-                var image = e.Item.ImageList.Images[e.Item.ImageIndex];
-                imageWidth = image.Width;
-                e.Graphics.DrawImageUnscaled(image, e.Item.Position.X, e.Item.Position.Y);
-            }
-
-            var font = FileStatusListView.Font;
-            var textStartX = e.Item.Position.X + imageWidth;
-            var textSpace = e.Item.Bounds.Width - textStartX;
-            var formatter = new PathFormatter(e.Graphics, font);
-
-            var text = formatter.FormatTextForDrawing(textSpace, gitItemStatus.Name, gitItemStatus.OldName);
-
-            text = AppendItemSubmoduleStatus(text, gitItemStatus);
-
-            var slashIndex = text.LastIndexOf('/');
-
-            var textRect = new Rectangle(textStartX, e.Item.Bounds.Top, textSpace, e.Item.Bounds.Height);
-
-            if (slashIndex == -1 || slashIndex >= text.Length - 1)
-            {
-                formatter.DrawString(text, textRect, SystemColors.ControlText);
-                return;
-            }
-
-            var prefix = text.Substring(0, slashIndex + 1);
-            var tail = text.Substring(slashIndex + 1);
-
-            var prefixSize = formatter.MeasureString(prefix);
-
-            DrawString(textRect, prefix, SystemColors.GrayText);
-
-            textRect.Offset(prefixSize.Width, 0);
-            DrawString(textRect, tail, SystemColors.ControlText);
-
-            void DrawString(Rectangle rect, string s, Color color)
-            {
-                rect.Intersect(Rectangle.Round(e.Graphics.ClipBounds));
-                if (rect.Width != 0 && rect.Height != 0)
+                ListViewItem sortedFirstGroupItem = FileStatusListView.Items.Cast<ListViewItem>().FirstOrDefault(item => item.Group == group);
+                if (sortedFirstGroupItem != null)
                 {
-                    formatter.DrawString(s, rect, color);
+                    sortedFirstGroupItem.Selected = true;
                 }
             }
-        }
-
-        private static string AppendItemSubmoduleStatus(string text, GitItemStatus item)
-        {
-            if (item.IsSubmodule &&
-                item.GetSubmoduleStatusAsync() != null &&
-                item.GetSubmoduleStatusAsync().IsCompleted &&
-                item.GetSubmoduleStatusAsync().CompletedResult() != null)
+            else
             {
-                text += item.GetSubmoduleStatusAsync().CompletedResult().AddedAndRemovedString();
+                FileStatusListView.Items[0].Selected = true;
             }
-
-            return text;
-        }
-
-        private void FileStatusListView_MouseDown(object sender, MouseEventArgs e)
-        {
-            // SELECT
-            if (e.Button == MouseButtons.Right)
-            {
-                var hover = FileStatusListView.HitTest(e.Location);
-
-                if (hover.Item != null && !hover.Item.Selected)
-                {
-                    ClearSelected();
-
-                    hover.Item.Selected = true;
-                }
-            }
-
-            // DRAG
-            if (e.Button == MouseButtons.Left)
-            {
-                if (SelectedItems.Any())
-                {
-                    // Remember the point where the mouse down occurred.
-                    // The DragSize indicates the size that the mouse can move
-                    // before a drag event should be started.
-                    Size dragSize = SystemInformation.DragSize;
-
-                    // Create a rectangle using the DragSize, with the mouse position being
-                    // at the center of the rectangle.
-                    _dragBoxFromMouseDown = new Rectangle(
-                        new Point(
-                            e.X - (dragSize.Width / 2),
-                            e.Y - (dragSize.Height / 2)),
-                        dragSize);
-                }
-                else
-                {
-                    // Reset the rectangle if the mouse is not over an item in the ListView.
-                    _dragBoxFromMouseDown = Rectangle.Empty;
-                }
-            }
-        }
-
-        public override ContextMenuStrip ContextMenuStrip
-        {
-            get { return FileStatusListView.ContextMenuStrip; }
-            set
-            {
-                FileStatusListView.ContextMenuStrip = value;
-                if (FileStatusListView.ContextMenuStrip != null)
-                {
-                    FileStatusListView.ContextMenuStrip.Opening += FileStatusListView_ContextMenu_Opening;
-                }
-            }
-        }
-
-        public override ContextMenu ContextMenu
-        {
-            get => FileStatusListView.ContextMenu;
-            set => FileStatusListView.ContextMenu = value;
-        }
-
-        private Rectangle _dragBoxFromMouseDown;
-
-        private void FileStatusListView_MouseMove(object sender, MouseEventArgs e)
-        {
-            ListView listView = sender as ListView;
-
-            // DRAG
-            // If the mouse moves outside the rectangle, start the drag.
-            if (_dragBoxFromMouseDown != Rectangle.Empty &&
-                !_dragBoxFromMouseDown.Contains(e.X, e.Y))
-            {
-                if (SelectedItems.Any())
-                {
-                    var fileList = new StringCollection();
-
-                    foreach (GitItemStatus item in SelectedItems)
-                    {
-                        string fileName = _fullPathResolver.Resolve(item.Name);
-
-                        fileList.Add(fileName.ToNativePath());
-                    }
-
-                    var obj = new DataObject();
-                    obj.SetFileDropList(fileList);
-
-                    // Proceed with the drag and drop, passing in the list item.
-                    DoDragDrop(obj, DragDropEffects.Copy);
-                    _dragBoxFromMouseDown = Rectangle.Empty;
-                }
-            }
-
-            // TOOLTIP
-            if (listView != null)
-            {
-                ListViewItem hoveredItem;
-                try
-                {
-                    var point = new Point(e.X, e.Y);
-                    hoveredItem = listView.HitTest(point).Item;
-                }
-                catch (ArgumentOutOfRangeException)
-                {
-                    hoveredItem = null;
-                }
-
-                if (hoveredItem?.Tag is GitItemStatus gitItemStatus)
-                {
-                    string text;
-                    if (gitItemStatus.IsRenamed || gitItemStatus.IsCopied)
-                    {
-                        text = string.Concat(gitItemStatus.Name, " (", gitItemStatus.OldName, ")");
-                    }
-                    else
-                    {
-                        text = gitItemStatus.Name;
-                    }
-
-                    float textWidth;
-                    using (var graphics = listView.CreateGraphics())
-                    {
-                        textWidth = graphics.MeasureString(text, listView.Font).Width + 17;
-                    }
-
-                    // Use width-itemheight because the icon drawn in front of the text is the itemheight
-                    if (textWidth > (FileStatusListView.Width - FileStatusListView.GetItemRect(hoveredItem.Index).Height))
-                    {
-                        if (hoveredItem.ToolTipText != gitItemStatus.ToString())
-                        {
-                            hoveredItem.ToolTipText = gitItemStatus.ToString();
-                        }
-                    }
-                    else
-                    {
-                        hoveredItem.ToolTipText = "";
-                    }
-                }
-            }
-        }
-
-        private void FileStatusListView_GroupMouseDown(object sender, ListViewGroupMouseEventArgs e)
-        {
-            if (e.Button == MouseButtons.Left)
-            {
-                // prevent selecting all sub-items when left-clicking group
-                e.Handled = true;
-            }
-        }
-
-        public int UnfilteredItemsCount => GitItemStatusesWithParents?.Sum(tuple => tuple.statuses.Count) ?? 0;
-
-        public int AllItemsCount => FileStatusListView.Items.Count;
-
-        [Browsable(false)]
-        public IEnumerable<GitItemStatus> AllItems => FileStatusListView.Items.Cast<ListViewItem>().Select(selectedItem => (GitItemStatus)selectedItem.Tag);
-
-        [Browsable(false)]
-        public IEnumerable<GitItemStatus> SelectedItems
-        {
-            get { return FileStatusListView.SelectedItems.Cast<ListViewItem>().Select(i => (GitItemStatus)i.Tag); }
-            set
-            {
-                ClearSelected();
-                if (value == null)
-                {
-                    return;
-                }
-
-                foreach (var item in FileStatusListView.Items.Cast<ListViewItem>()
-                    .Where(i => value.Contains(i.Tag as GitItemStatus)))
-                {
-                    item.Selected = true;
-                }
-
-                var first = FileStatusListView.SelectedItems.Cast<ListViewItem>().FirstOrDefault(x => x.Selected);
-                first?.EnsureVisible();
-                StoreNextIndexToSelect();
-            }
-        }
-
-        [CanBeNull]
-        [DesignerSerializationVisibility(DesignerSerializationVisibility.Hidden)]
-        [Browsable(false)]
-        public GitItemStatus SelectedItem
-        {
-            get { return SelectedItems.FirstOrDefault(); }
-            set
-            {
-                ClearSelected();
-                if (value == null)
-                {
-                    return;
-                }
-
-                ListViewItem newSelected = null;
-                foreach (ListViewItem item in FileStatusListView.Items)
-                {
-                    if (value.CompareTo(item.Tag as GitItemStatus) == 0)
-                    {
-                        if (newSelected == null)
-                        {
-                            newSelected = item;
-                        }
-                        else if (item.Tag == value)
-                        {
-                            newSelected = item;
-                            break;
-                        }
-                    }
-                }
-
-                if (newSelected != null)
-                {
-                    newSelected.Selected = true;
-                    newSelected.EnsureVisible();
-                }
-            }
-        }
-
-        [CanBeNull]
-        [DesignerSerializationVisibility(DesignerSerializationVisibility.Hidden)]
-        [Browsable(false)]
-        public GitRevision SelectedItemParent => SelectedItemParents.FirstOrDefault();
-
-        [DesignerSerializationVisibility(DesignerSerializationVisibility.Hidden)]
-        [Browsable(false)]
-        public IEnumerable<GitRevision> SelectedItemParents
-        {
-            get
-            {
-                return FileStatusListView.SelectedItems.Cast<ListViewItem>()
-                    .Where(i => i.Group?.Tag is GitRevision)
-                    .Select(i => (GitRevision)i.Group.Tag);
-            }
-        }
-
-        [DesignerSerializationVisibility(DesignerSerializationVisibility.Hidden)]
-        [Browsable(false)]
-        public IEnumerable<GitItemStatusWithParent> SelectedItemsWithParent
-        {
-            get
-            {
-                return FileStatusListView.SelectedItems.Cast<ListViewItem>()
-                    .Where(i => i.Group?.Tag is GitRevision)
-                    .Select(i => new GitItemStatusWithParent((GitRevision)i.Group.Tag, i.Tag as GitItemStatus));
-            }
-        }
-
-        public void ClearSelected()
-        {
-            foreach (ListViewItem item in FileStatusListView.SelectedItems)
-            {
-                item.Selected = false;
-            }
-        }
-
-        [DesignerSerializationVisibility(DesignerSerializationVisibility.Hidden)]
-        [Browsable(false)]
-        public int SelectedIndex
-        {
-            get
-            {
-                foreach (int i in FileStatusListView.SelectedIndices)
-                {
-                    return i;
-                }
-
-                return -1;
-            }
-            set
-            {
-                ClearSelected();
-                if (value >= 0)
-                {
-                    FileStatusListView.Items[value].Selected = true;
-                    FileStatusListView.Items[value].Focused = true;
-                    FileStatusListView.Items[value].EnsureVisible();
-                }
-            }
-        }
-
-        private int _nextIndexToSelect = -1;
-
-        public void StoreNextIndexToSelect()
-        {
-            _nextIndexToSelect = -1;
-            foreach (int idx in FileStatusListView.SelectedIndices)
-            {
-                if (idx > _nextIndexToSelect)
-                {
-                    _nextIndexToSelect = idx;
-                }
-            }
-
-            _nextIndexToSelect = _nextIndexToSelect - FileStatusListView.SelectedIndices.Count + 1;
         }
 
         public void SelectStoredNextIndex(int defaultIndex = -1)
@@ -681,29 +551,160 @@ namespace GitUI
             _nextIndexToSelect = -1;
         }
 
-        public event EventHandler SelectedIndexChanged;
-        public event EventHandler DataSourceChanged;
-
-        public new event EventHandler DoubleClick;
-        public new event KeyEventHandler KeyDown;
-
-        private void FileStatusListView_DoubleClick(object sender, EventArgs e)
+        public void SetDiffs(IReadOnlyList<GitRevision> revisions)
         {
-            if (DoubleClick == null)
+            Revision = revisions.FirstOrDefault();
+
+            var tuples = new List<(GitRevision revision, IReadOnlyList<GitItemStatus> statuses)>();
+
+            if (Revision != null)
             {
-                if (AppSettings.OpenSubmoduleDiffInSeparateWindow && SelectedItem.IsSubmodule)
+                GitRevision[] parentRevs;
+                if (revisions.Count == 1)
                 {
-                    ThreadHelper.JoinableTaskFactory.RunAsync(() => OpenSubmoduleAsync());
+                    // Note: RevisionGrid could in some forms be used to get the parent guids
+                    parentRevs = Revision.ParentIds?.Select(item => new GitRevision(item)).ToArray();
                 }
                 else
                 {
-                    UICommands.StartFileHistoryDialog(this, SelectedItem.Name, Revision);
+                    parentRevs = revisions.Skip(1).ToArray();
+                }
+
+                if (parentRevs == null || parentRevs.Length == 0)
+                {
+                    // No parent, will set null as parent
+                    tuples.Add((null, Module.GetTreeFiles(Revision.TreeGuid, full: true)));
+                }
+                else
+                {
+                    if (!AppSettings.ShowDiffForAllParents)
+                    {
+                        parentRevs = new[] { parentRevs[0] };
+                    }
+
+                    foreach (var rev in parentRevs)
+                    {
+                        tuples.Add((rev, Module.GetDiffFilesWithSubmodulesStatus(rev.Guid, Revision.Guid, Revision.ParentIds?.FirstOrDefault()?.ToString())));
+                    }
+
+                    // Show combined (merge conflicts) only when all first (A) are parents to selected (B)
+                    var isMergeCommit = AppSettings.ShowDiffForAllParents &&
+                                        Revision.ParentIds != null && Revision.ParentIds.Count > 1 &&
+                                        _revisionTester.AllFirstAreParentsToSelected(parentRevs, Revision);
+                    if (isMergeCommit)
+                    {
+                        var conflicts = Module.GetCombinedDiffFileList(Revision.Guid);
+                        if (conflicts.Count != 0)
+                        {
+                            // Create an artificial commit
+                            tuples.Add((new GitRevision(ObjectId.CombinedDiffId), conflicts));
+                        }
+                    }
                 }
             }
-            else
+
+            GitItemStatusesWithParents = tuples;
+        }
+
+        public void SetDiffs(GitRevision selectedRev = null, GitRevision parentRev = null, IReadOnlyList<GitItemStatus> items = null)
+        {
+            Revision = selectedRev;
+            GitItemStatusesWithParents = items == null
+                ? Array.Empty<(GitRevision, IReadOnlyList<GitItemStatus>)>()
+                : new[] { (parentRev, items) };
+        }
+
+        public void SetNoFilesText(string text)
+        {
+            NoFiles.Text = text;
+        }
+
+        public void SetSelectedIndex(int idx, bool notify)
+        {
+            _enableSelectedIndexChangeEvent = notify;
+            SelectedIndex = idx;
+            _enableSelectedIndexChangeEvent = true;
+        }
+
+        public int SetSelectionFilter(string selectionFilter)
+        {
+            return SelectFiles(RegexForSelecting(selectionFilter));
+
+            int SelectFiles(Regex regex)
             {
-                DoubleClick(sender, e);
+                try
+                {
+                    SuspendLayout();
+
+                    var i = 0;
+                    foreach (var item in AllItems)
+                    {
+                        FileStatusListView.Items[i].Selected = regex.IsMatch(item.Name);
+                        i++;
+                    }
+
+                    return FileStatusListView.SelectedIndices.Count;
+                }
+                finally
+                {
+                    ResumeLayout(true);
+                }
             }
+
+            Regex RegexForSelecting(string value)
+            {
+                return string.IsNullOrEmpty(value)
+                    ? new Regex("^$", RegexOptions.Compiled)
+                    : new Regex(value, RegexOptions.Compiled | RegexOptions.IgnoreCase);
+            }
+        }
+
+        public void StoreNextIndexToSelect()
+        {
+            _nextIndexToSelect = -1;
+            foreach (int idx in FileStatusListView.SelectedIndices)
+            {
+                if (idx > _nextIndexToSelect)
+                {
+                    _nextIndexToSelect = idx;
+                }
+            }
+
+            _nextIndexToSelect = _nextIndexToSelect - FileStatusListView.SelectedIndices.Count + 1;
+        }
+
+        // Protected methods
+
+        protected override void DisposeCustomResources()
+        {
+            _selectedIndexChangeSubscription?.Dispose();
+        }
+
+        // Private methods
+
+        private void CreateOpenSubmoduleMenuItem()
+        {
+            _openSubmoduleMenuItem = new ToolStripMenuItem
+            {
+                Name = "openSubmoduleMenuItem",
+                Tag = "1",
+                Text = "Open with Git Extensions",
+                Image = Images.GitExtensionsLogo16
+            };
+            _openSubmoduleMenuItem.Click += (s, ea) => { ThreadHelper.JoinableTaskFactory.RunAsync(() => OpenSubmoduleAsync()); };
+        }
+
+        private static string AppendItemSubmoduleStatus(string text, GitItemStatus item)
+        {
+            if (item.IsSubmodule &&
+                item.GetSubmoduleStatusAsync() != null &&
+                item.GetSubmoduleStatusAsync().IsCompleted &&
+                item.GetSubmoduleStatusAsync().CompletedResult() != null)
+            {
+                text += item.GetSubmoduleStatusAsync().CompletedResult().AddedAndRemovedString();
+            }
+
+            return text;
         }
 
         private async Task OpenSubmoduleAsync()
@@ -723,79 +724,6 @@ namespace GitUI
             };
 
             process.Start();
-        }
-
-        private void FileStatusListView_ContextMenu_Opening(object sender, CancelEventArgs e)
-        {
-            var cm = (ContextMenuStrip)sender;
-
-            if (!cm.Items.Find(_openSubmoduleMenuItem.Name, true).Any())
-            {
-                cm.Items.Insert(1, _openSubmoduleMenuItem);
-            }
-
-            bool isSubmoduleSelected = SelectedItem != null && SelectedItem.IsSubmodule;
-
-            _openSubmoduleMenuItem.Visible = isSubmoduleSelected;
-
-            if (isSubmoduleSelected)
-            {
-                _openSubmoduleMenuItem.Font = AppSettings.OpenSubmoduleDiffInSeparateWindow ?
-                    new Font(_openSubmoduleMenuItem.Font, FontStyle.Bold) :
-                    new Font(_openSubmoduleMenuItem.Font, FontStyle.Regular);
-            }
-        }
-
-        private void FileStatusListView_SelectedIndexChanged()
-        {
-            SelectedIndexChanged?.Invoke(this, EventArgs.Empty);
-        }
-
-        [Browsable(false)]
-        [DefaultValue(true)]
-        public bool IsEmpty => GitItemStatuses == null || !GitItemStatuses.Any();
-
-        [DesignerSerializationVisibility(DesignerSerializationVisibility.Hidden)]
-        [Browsable(false)]
-        public IReadOnlyList<GitItemStatus> GitItemStatuses
-        {
-            get
-            {
-                return GitItemStatusesWithParents?.SelectMany(tuple => tuple.statuses).ToList()
-                       ?? (IReadOnlyList<GitItemStatus>)Array.Empty<GitItemStatus>();
-            }
-        }
-
-        [DesignerSerializationVisibility(DesignerSerializationVisibility.Hidden)]
-        [Browsable(false)]
-        public IReadOnlyList<GitItemStatus> GitItemFilteredStatuses
-        {
-            get
-            {
-                var result = new List<GitItemStatus>(FileStatusListView.Items.Count);
-
-                foreach (ListViewItem listViewItem in FileStatusListView.Items)
-                {
-                    result.Add(listViewItem.Tag as GitItemStatus);
-                }
-
-                return result;
-            }
-        }
-
-        private IReadOnlyList<(GitRevision revision, IReadOnlyList<GitItemStatus> statuses)> _itemsWithParent = Array.Empty<(GitRevision, IReadOnlyList<GitItemStatus>)>();
-
-        // Parents is used as the "first selected" (not always the parent) for GitItemStatus
-        [DesignerSerializationVisibility(DesignerSerializationVisibility.Hidden)]
-        [Browsable(false)]
-        public IReadOnlyList<(GitRevision revision, IReadOnlyList<GitItemStatus> statuses)> GitItemStatusesWithParents
-        {
-            get { return _itemsWithParent; }
-            private set
-            {
-                _itemsWithParent = value ?? throw new ArgumentNullException(nameof(value));
-                UpdateFileStatusListView();
-            }
         }
 
         private void UpdateFileStatusListView(bool updateCausedByFilter = false)
@@ -824,7 +752,7 @@ namespace GitUI
             }
 
             FileStatusListView.BeginUpdate();
-            FileStatusListView.ShowGroups = GitItemStatusesWithParents.Count > 1 || _alwaysRevisionGroups;
+            FileStatusListView.ShowGroups = GitItemStatusesWithParents.Count > 1 || _groupByRevision;
             FileStatusListView.Groups.Clear();
             FileStatusListView.Items.Clear();
 
@@ -1020,192 +948,6 @@ namespace GitUI
             }
         }
 
-        [DefaultValue(true)]
-        public bool SelectFirstItemOnSetItems { get; set; }
-
-        public void SelectFirstVisibleItem()
-        {
-            if (FileStatusListView.Items.Count == 0)
-            {
-                return;
-            }
-
-            var group = FileStatusListView.Groups.Cast<ListViewGroup>().FirstOrDefault(gr => gr.Items.Count > 0);
-            if (group != null)
-            {
-                ListViewItem sortedFirstGroupItem = FileStatusListView.Items.Cast<ListViewItem>().FirstOrDefault(item => item.Group == group);
-                if (sortedFirstGroupItem != null)
-                {
-                    sortedFirstGroupItem.Selected = true;
-                }
-            }
-            else
-            {
-                FileStatusListView.Items[0].Selected = true;
-            }
-        }
-
-        /// <summary>
-        /// Gets or sets the revision.
-        /// </summary>
-        /// <value>The revision.</value>
-        [DesignerSerializationVisibility(DesignerSerializationVisibility.Hidden)]
-        [Browsable(false)]
-        public GitRevision Revision { get; set; }
-
-        private void FileStatusListView_KeyDown(object sender, KeyEventArgs e)
-        {
-            switch (e.KeyCode)
-            {
-                case Keys.A:
-                {
-                    if (!e.Control)
-                    {
-                        break;
-                    }
-
-                    FileStatusListView.BeginUpdate();
-                    try
-                    {
-                        for (var i = 0; i < FileStatusListView.Items.Count; i++)
-                        {
-                            FileStatusListView.Items[i].Selected = true;
-                        }
-
-                        e.Handled = true;
-                    }
-                    finally
-                    {
-                        FileStatusListView.EndUpdate();
-                    }
-
-                    break;
-                }
-
-                default:
-                {
-                    KeyDown?.Invoke(sender, e);
-                    break;
-                }
-            }
-        }
-
-        public int SetSelectionFilter(string selectionFilter)
-        {
-            return SelectFiles(RegexForSelecting(selectionFilter));
-
-            int SelectFiles(Regex regex)
-            {
-                try
-                {
-                    SuspendLayout();
-
-                    var i = 0;
-                    foreach (var item in AllItems)
-                    {
-                        FileStatusListView.Items[i].Selected = regex.IsMatch(item.Name);
-                        i++;
-                    }
-
-                    return FileStatusListView.SelectedIndices.Count;
-                }
-                finally
-                {
-                    ResumeLayout(true);
-                }
-            }
-
-            Regex RegexForSelecting(string value)
-            {
-                return string.IsNullOrEmpty(value)
-                    ? new Regex("^$", RegexOptions.Compiled)
-                    : new Regex(value, RegexOptions.Compiled | RegexOptions.IgnoreCase);
-            }
-        }
-
-        public void SelectAll()
-        {
-            try
-            {
-                SuspendLayout();
-
-                var itemCount = AllItems.Count();
-
-                for (var i = 0; i < itemCount; i++)
-                {
-                    FileStatusListView.Items[i].Selected = true;
-                    i++;
-                }
-            }
-            finally
-            {
-                ResumeLayout(true);
-            }
-        }
-
-        public void SetDiffs(GitRevision selectedRev = null, GitRevision parentRev = null, IReadOnlyList<GitItemStatus> items = null)
-        {
-            Revision = selectedRev;
-            GitItemStatusesWithParents = items == null
-                ? Array.Empty<(GitRevision, IReadOnlyList<GitItemStatus>)>()
-                : new[] { (parentRev, items) };
-        }
-
-        public void SetDiffs(IReadOnlyList<GitRevision> revisions)
-        {
-            Revision = revisions.FirstOrDefault();
-
-            var tuples = new List<(GitRevision revision, IReadOnlyList<GitItemStatus> statuses)>();
-
-            if (Revision != null)
-            {
-                GitRevision[] parentRevs;
-                if (revisions.Count == 1)
-                {
-                    // Note: RevisionGrid could in some forms be used to get the parent guids
-                    parentRevs = Revision.ParentIds?.Select(item => new GitRevision(item)).ToArray();
-                }
-                else
-                {
-                    parentRevs = revisions.Skip(1).ToArray();
-                }
-
-                if (parentRevs == null || parentRevs.Length == 0)
-                {
-                    // No parent, will set null as parent
-                    tuples.Add((null, Module.GetTreeFiles(Revision.TreeGuid, full: true)));
-                }
-                else
-                {
-                    if (!AppSettings.ShowDiffForAllParents)
-                    {
-                        parentRevs = new[] { parentRevs[0] };
-                    }
-
-                    foreach (var rev in parentRevs)
-                    {
-                        tuples.Add((rev, Module.GetDiffFilesWithSubmodulesStatus(rev.Guid, Revision.Guid, Revision.ParentIds?.FirstOrDefault()?.ToString())));
-                    }
-
-                    // Show combined (merge conflicts) only when all first (A) are parents to selected (B)
-                    var isMergeCommit = AppSettings.ShowDiffForAllParents &&
-                                        Revision.ParentIds != null && Revision.ParentIds.Count > 1 &&
-                                        _revisionTester.AllFirstAreParentsToSelected(parentRevs, Revision);
-                    if (isMergeCommit)
-                    {
-                        var conflicts = Module.GetCombinedDiffFileList(Revision.Guid);
-                        if (conflicts.Count != 0)
-                        {
-                            // Create an artificial commit
-                            tuples.Add((new GitRevision(ObjectId.CombinedDiffId), conflicts));
-                        }
-                    }
-                }
-            }
-
-            GitItemStatusesWithParents = tuples;
-        }
-
         private void HandleVisibility_NoFilesLabel_FilterComboBox(bool filesPresent)
         {
             NoFiles.Visible = !filesPresent;
@@ -1213,6 +955,271 @@ namespace GitUI
             {
                 FilterComboBox.Visible = filesPresent;
             }
+        }
+
+        // Event handlers
+
+        private void FileStatusListView_ContextMenu_Opening(object sender, CancelEventArgs e)
+        {
+            var cm = (ContextMenuStrip)sender;
+
+            if (!cm.Items.Find(_openSubmoduleMenuItem.Name, true).Any())
+            {
+                cm.Items.Insert(1, _openSubmoduleMenuItem);
+            }
+
+            bool isSubmoduleSelected = SelectedItem != null && SelectedItem.IsSubmodule;
+
+            _openSubmoduleMenuItem.Visible = isSubmoduleSelected;
+
+            if (isSubmoduleSelected)
+            {
+                _openSubmoduleMenuItem.Font = AppSettings.OpenSubmoduleDiffInSeparateWindow ?
+                    new Font(_openSubmoduleMenuItem.Font, FontStyle.Bold) :
+                    new Font(_openSubmoduleMenuItem.Font, FontStyle.Regular);
+            }
+        }
+
+        private void FileStatusListView_DoubleClick(object sender, EventArgs e)
+        {
+            if (DoubleClick == null)
+            {
+                if (AppSettings.OpenSubmoduleDiffInSeparateWindow && SelectedItem.IsSubmodule)
+                {
+                    ThreadHelper.JoinableTaskFactory.RunAsync(() => OpenSubmoduleAsync());
+                }
+                else
+                {
+                    UICommands.StartFileHistoryDialog(this, SelectedItem.Name, Revision);
+                }
+            }
+            else
+            {
+                DoubleClick(sender, e);
+            }
+        }
+
+        private void FileStatusListView_DrawSubItem(object sender, DrawListViewSubItemEventArgs e)
+        {
+            if (!(e.Item?.Tag is GitItemStatus gitItemStatus))
+            {
+                return;
+            }
+
+            var imageWidth = 0;
+            if (e.Item.ImageList != null && e.Item.ImageIndex != -1)
+            {
+                var image = e.Item.ImageList.Images[e.Item.ImageIndex];
+                imageWidth = image.Width;
+                e.Graphics.DrawImageUnscaled(image, e.Item.Position.X, e.Item.Position.Y);
+            }
+
+            var font = FileStatusListView.Font;
+            var textStartX = e.Item.Position.X + imageWidth;
+            var textSpace = e.Item.Bounds.Width - textStartX;
+            var formatter = new PathFormatter(e.Graphics, font);
+
+            var text = formatter.FormatTextForDrawing(textSpace, gitItemStatus.Name, gitItemStatus.OldName);
+
+            text = AppendItemSubmoduleStatus(text, gitItemStatus);
+
+            var slashIndex = text.LastIndexOf('/');
+
+            var textRect = new Rectangle(textStartX, e.Item.Bounds.Top, textSpace, e.Item.Bounds.Height);
+
+            if (slashIndex == -1 || slashIndex >= text.Length - 1)
+            {
+                formatter.DrawString(text, textRect, SystemColors.ControlText);
+                return;
+            }
+
+            var prefix = text.Substring(0, slashIndex + 1);
+            var tail = text.Substring(slashIndex + 1);
+
+            var prefixSize = formatter.MeasureString(prefix);
+
+            DrawString(textRect, prefix, SystemColors.GrayText);
+
+            textRect.Offset(prefixSize.Width, 0);
+            DrawString(textRect, tail, SystemColors.ControlText);
+
+            void DrawString(Rectangle rect, string s, Color color)
+            {
+                rect.Intersect(Rectangle.Round(e.Graphics.ClipBounds));
+                if (rect.Width != 0 && rect.Height != 0)
+                {
+                    formatter.DrawString(s, rect, color);
+                }
+            }
+        }
+
+        private void FileStatusListView_GroupMouseDown(object sender, ListViewGroupMouseEventArgs e)
+        {
+            if (e.Button == MouseButtons.Left)
+            {
+                // prevent selecting all sub-items when left-clicking group
+                e.Handled = true;
+            }
+        }
+
+        private void FileStatusListView_KeyDown(object sender, KeyEventArgs e)
+        {
+            switch (e.KeyCode)
+            {
+                case Keys.A:
+                    {
+                        if (!e.Control)
+                        {
+                            break;
+                        }
+
+                        FileStatusListView.BeginUpdate();
+                        try
+                        {
+                            for (var i = 0; i < FileStatusListView.Items.Count; i++)
+                            {
+                                FileStatusListView.Items[i].Selected = true;
+                            }
+
+                            e.Handled = true;
+                        }
+                        finally
+                        {
+                            FileStatusListView.EndUpdate();
+                        }
+
+                        break;
+                    }
+
+                default:
+                    {
+                        KeyDown?.Invoke(sender, e);
+                        break;
+                    }
+            }
+        }
+
+        private void FileStatusListView_MouseDown(object sender, MouseEventArgs e)
+        {
+            // SELECT
+            if (e.Button == MouseButtons.Right)
+            {
+                var hover = FileStatusListView.HitTest(e.Location);
+
+                if (hover.Item != null && !hover.Item.Selected)
+                {
+                    ClearSelected();
+
+                    hover.Item.Selected = true;
+                }
+            }
+
+            // DRAG
+            if (e.Button == MouseButtons.Left)
+            {
+                if (SelectedItems.Any())
+                {
+                    // Remember the point where the mouse down occurred.
+                    // The DragSize indicates the size that the mouse can move
+                    // before a drag event should be started.
+                    Size dragSize = SystemInformation.DragSize;
+
+                    // Create a rectangle using the DragSize, with the mouse position being
+                    // at the center of the rectangle.
+                    _dragBoxFromMouseDown = new Rectangle(
+                        new Point(
+                            e.X - (dragSize.Width / 2),
+                            e.Y - (dragSize.Height / 2)),
+                        dragSize);
+                }
+                else
+                {
+                    // Reset the rectangle if the mouse is not over an item in the ListView.
+                    _dragBoxFromMouseDown = Rectangle.Empty;
+                }
+            }
+        }
+
+        private void FileStatusListView_MouseMove(object sender, MouseEventArgs e)
+        {
+            ListView listView = sender as ListView;
+
+            // DRAG
+            // If the mouse moves outside the rectangle, start the drag.
+            if (_dragBoxFromMouseDown != Rectangle.Empty &&
+                !_dragBoxFromMouseDown.Contains(e.X, e.Y))
+            {
+                if (SelectedItems.Any())
+                {
+                    var fileList = new StringCollection();
+
+                    foreach (GitItemStatus item in SelectedItems)
+                    {
+                        string fileName = _fullPathResolver.Resolve(item.Name);
+
+                        fileList.Add(fileName.ToNativePath());
+                    }
+
+                    var obj = new DataObject();
+                    obj.SetFileDropList(fileList);
+
+                    // Proceed with the drag and drop, passing in the list item.
+                    DoDragDrop(obj, DragDropEffects.Copy);
+                    _dragBoxFromMouseDown = Rectangle.Empty;
+                }
+            }
+
+            // TOOLTIP
+            if (listView != null)
+            {
+                ListViewItem hoveredItem;
+                try
+                {
+                    var point = new Point(e.X, e.Y);
+                    hoveredItem = listView.HitTest(point).Item;
+                }
+                catch (ArgumentOutOfRangeException)
+                {
+                    hoveredItem = null;
+                }
+
+                if (hoveredItem?.Tag is GitItemStatus gitItemStatus)
+                {
+                    string text;
+                    if (gitItemStatus.IsRenamed || gitItemStatus.IsCopied)
+                    {
+                        text = string.Concat(gitItemStatus.Name, " (", gitItemStatus.OldName, ")");
+                    }
+                    else
+                    {
+                        text = gitItemStatus.Name;
+                    }
+
+                    float textWidth;
+                    using (var graphics = listView.CreateGraphics())
+                    {
+                        textWidth = graphics.MeasureString(text, listView.Font).Width + 17;
+                    }
+
+                    // Use width-itemheight because the icon drawn in front of the text is the itemheight
+                    if (textWidth > (FileStatusListView.Width - FileStatusListView.GetItemRect(hoveredItem.Index).Height))
+                    {
+                        if (hoveredItem.ToolTipText != gitItemStatus.ToString())
+                        {
+                            hoveredItem.ToolTipText = gitItemStatus.ToString();
+                        }
+                    }
+                    else
+                    {
+                        hoveredItem.ToolTipText = "";
+                    }
+                }
+            }
+        }
+
+        private void FileStatusListView_SelectedIndexChanged()
+        {
+            SelectedIndexChanged?.Invoke(this, EventArgs.Empty);
         }
 
         #region Filtering
@@ -1316,17 +1323,5 @@ namespace GitUI
         }
 
         #endregion
-    }
-
-    public sealed class GitItemStatusWithParent
-    {
-        public GitRevision ParentRevision { get; }
-        public GitItemStatus Item { get; }
-
-        public GitItemStatusWithParent(GitRevision parent, GitItemStatus item)
-        {
-            ParentRevision = parent;
-            Item = item;
-        }
     }
 }
